User can choose spots and take turns in the game
Now the user has the ability to select the location of the board, place their O or X, and switch the turns. Also added further data validation to check if there is an X or O already taken on the board, a win condition has yet to be added for now.