Fully function tic tac toe game for two people made
I have added the checkForWin function which checks if there is a 3 in a row on the board. Also in the main function,  I have added some conditions to check that if one of the players win, ask them if they would like to play again, if so reset the board and begin the game again.