Added option to play against CPU
This change now adds the option to let the user choose between playing with a player next to them, or play against an AI. I have used the minimax algorithm for the AI logic to play against the human player, and the AI is able to successfully choose the best possible moves given the situation and block the human player from winning as well, all while making valid moves.

Also, I have optimized some existing functions and added some new functions in the code from the previous commit. like a function that resets the game board or switches the player when it is called.