Optimized the AI and Added Alpha-Beta Pruning
First I will begin with the minimax algorithm. I notice that when playing against the AI it will still choose the move that will guarantee it a win, however it will not always pick the move that will give them that win in the least amount of turns. Now I have it where the AI will always pick the best possible move against the human player using the least amount of turns possible. To fix this I simply subtracted the depth from the return under maximizing, and added depth under the minimizing if statement in the miniMax function.

I have also added alpha-beta pruning to the miniMax algorithm in order to make the algorithm more efficient and search for only moves that will lead it to a win, and prune other moves that will end up with a similar or undesirable result for the AI.